requirements.txt
discord.py[voice]
jishaku
